Get scenarios from API and display in app
