--- client/components/app/app.css
@@ -1,4 +1,0 @@
-.tm__app {
-    margin: 1rem;
-    font-family: Arial, Helvetica, sans-serif;
-}
