+++ client/static/index.html
@@ -4,7 +4,7 @@
     <title>Teacher Moments</title>
 </head>
 <body>
-    <div id="root"></div>
+    <div id="root" class="tm__app"></div>
 
     <script src="bundle.js"></script>
 </body>
